Update twitch functionality
This script was previously using the Twitch v5 API "Kraken". This version of the API is effectively not available to any applications younger than July 15, 2021. See https://blog.twitch.tv/en/2021/07/15/legacy-twitch-api-v5-shutdown-details-and-timeline/?utm_referrer=https://dev.twitch.tv/. Of course, I didn't realize that when I modeled the Twitch functionality after an older example bot...
The new Twitch API is called "Helix". Many endpoints have been changed; the format of data returned from requests has changed; and applications require oAuth tokens for all requests, from what I can tell.
This all means that the Twitch functionality of this script needed a huge rewrite.
This version of the commit is still not finished and there's a lot of 'junk' for debugging purposes, but it can now successfully check whether tactix11b is streaming or not, using the debug command !checktwitch.