Add Twitch token refreshing
The bot was running just fine for around 10 days, then stopped announcing streams properly.

The bot previously did not store the Twitch "refresh" token, which is used to extend the lifetime of the current OAuth token. In this update, the bot now saves the refresh token, and refreshes the OAuth token as needed, instead of immediately asking for brand new tokens when seeing a 401 (not authorized) error.

Added a check_token discord command, which tells how many seconds remain until the current OAuth token expires.

Updated formatting.